[GLOBAL] Cleanup `metadata.rb` for better maintainability
This introduces the following changes:

- Add missing `metadata` to `metadata.rb`
- Fix dependencies, indentions and style to make them easier to read and
maintain.

The idea is to use `metadata.rb` as a single source of truth for our
documentation. Thus it is very important that it can be maintained
easily and in a consistent manner.

This is a first step into this direction.

See http://docs.opscode.com/config_rb_metadata.html for details on this
file.