[GLOBAL] Add autogenerated `README.md` to each cookbook
In order to document our own / patched cookbooks with a README we add
them now.

Sadly the README does not combine the 'original' infos from `cookbooks`
for patched cookbooks.

We also enrich the comments in the source code to make them 'visible'
in the README.md file with the `#<...#>` syntax.

See https://github.com/realityforge/knife-cookbook-doc#getting-started
for more context.

This is an easy way to populate the README with 'up-to-date' content.

Use the source, Luke :)